requirements.txt
# Core dependencies for VersionTracker
pyyaml>=6.0
click>=8.0
tqdm>=4.65
psutil>=5.9

# Optional but recommended
rapidfuzz>=3.0
# Updated for PR #32 security compliance
